IPv6 support.
* grecs: Upgrade.
* configure.ac (GRECS_SETUP): Request sockaddr-list.

* include/dico/types.h (DICO_DICT_PORT_STR): New constant.
(dico_url) <port>: Change type to char *. All uses changed.
* include/dico/url.h (ip_hostname, get_ipaddr)
(str2port): Remove obsolete functions and data types.
* include/appi18n.h (appi18n_init): New proto.

* dicod/accesslog.c: Add typecasts to struct sockaddr*. Fix
format string conversions.
* dicod/acl.c (acl_value): New union.
(acl_entry) <sockaddrs>: Rename to values. Holds acl_value entries now.
(_parse_sockaddr): Remove.
(_parse_acl_value): New function.  Use grecs CIDRs to represent ACL entries.
(_check_sockaddr): Replace with _check_value. All uses updated.
* dicod/alias.c (alias_hasher): Fix signature.
* dicod/dicod.c: Add typecasts to struct sockaddr*.
* dicod/dicod.h (client_addr): Change type to sockaddr_storage.
* dicod/ident.c (query_ident_name): Remove unused variable.
* dicod/loader.c: Minor fix.
* dicod/main.c (cb_dico_sockaddr_list): Use grecs sockaddr-list support.
* dicod/server.c (address_family_to_domain): Add IPv6 support.
(client_addr): Change type to struct sockaddr_storage. All uses changed.

* dico/cmdline.opt: Update.
* dico/connect.c (dict_connect): Rewrite using getaddrinfo.
* dico/dico-priv.h (source_addr): Change type to char*.
(shell_banner): New proto.
* dico/dico.c (source_addr): Change type to char*.
(fixup_url): dico_url.port member is char *.
* dico/func.c: dico_url.port member is char *.

* lib/iputil.c (ip_hostname, get_ipaddr)
(str2port): Remove obsolete functions.
(sockaddr_to_str): Add IPv6 support.
* lib/url.c (url_get_host): Allocate memory for url.port
(dico_url_destroy): Free port.

* lib/logstream.c: Minor fixes.
* lib/qp.c: Likewise.
* modules/gcide/markup.l: Likewise.

+++ dicod/alias.c
@@ -28,7 +28,7 @@ static Hash_table *alias_table;
 
 /* Calculate the hash of a struct input_file_ident.  */
 static size_t
-alias_hasher(void const *data, unsigned n_buckets)
+alias_hasher(void const *data, size_t n_buckets)
 {
     const struct alias *ap = data;
     return hash_string(ap->kw, n_buckets);
